@@ -1,1 +1,1 @@
-ruby-2.2
+ruby-2.3
