@@ -18,7 +18,7 @@ shared_examples_for "a schema based apartment adapter" do
     it "should process model exclusions" do
       Apartment::Tenant.init
 
-      Company.table_name.should == "public.companies"
+      expect(Company.table_name).to eq("public.companies")
     end
 
     context "with a default_schema", :default_schema => true do
@@ -26,20 +26,20 @@ shared_examples_for "a schema based apartment adapter" do
       it "should set the proper table_name on excluded_models" do
         Apartment::Tenant.init
 
-        Company.table_name.should == "#{default_schema}.companies"
+        expect(Company.table_name).to eq("#{default_schema}.companies")
       end
 
       it 'sets the search_path correctly' do
         Apartment::Tenant.init
 
-        User.connection.schema_search_path.should =~ %r|#{default_schema}|
+        expect(User.connection.schema_search_path).to match(%r|#{default_schema}|)
       end
     end
 
     context "persistent_schemas", :persistent_schemas => true do
       it "sets the persistent schemas in the schema_search_path" do
         Apartment::Tenant.init
-        connection.schema_search_path.should end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
+        expect(connection.schema_search_path).to end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
       end
     end
   end
@@ -51,7 +51,7 @@ shared_examples_for "a schema based apartment adapter" do
 
     it "should load schema.rb to new schema" do
       connection.schema_search_path = schema1
-      connection.tables.should include('companies')
+      expect(connection.tables).to include('companies')
     end
 
     it "should yield to block if passed and reset" do
@@ -61,13 +61,13 @@ shared_examples_for "a schema based apartment adapter" do
 
       subject.create(schema2) do
         @count = User.count
-        connection.schema_search_path.should start_with %{"#{schema2}"}
+        expect(connection.schema_search_path).to start_with %{"#{schema2}"}
         User.create
       end
 
-      connection.schema_search_path.should_not start_with %{"#{schema2}"}
+      expect(connection.schema_search_path).not_to start_with %{"#{schema2}"}
 
-      subject.switch(schema2){ User.count.should == @count + 1 }
+      subject.switch(schema2){ expect(User.count).to eq(@count + 1) }
     end
 
     context "numeric database names" do
@@ -76,7 +76,7 @@ shared_examples_for "a schema based apartment adapter" do
         expect {
           subject.create(db)
         }.to_not raise_error
-        tenant_names.should include(db.to_s)
+        expect(tenant_names).to include(db.to_s)
       end
 
       after{ subject.drop(db) }
@@ -99,7 +99,7 @@ shared_examples_for "a schema based apartment adapter" do
         expect {
           subject.drop(db)
         }.to_not raise_error
-        tenant_names.should_not include(db.to_s)
+        expect(tenant_names).not_to include(db.to_s)
       end
 
       after { subject.drop(db) rescue nil }
@@ -109,10 +109,10 @@ shared_examples_for "a schema based apartment adapter" do
   describe "#switch" do
     it "connects and resets" do
       subject.switch(schema1) do
-        connection.schema_search_path.should start_with %{"#{schema1}"}
+        expect(connection.schema_search_path).to start_with %{"#{schema1}"}
       end
 
-      connection.schema_search_path.should start_with %{"#{public_schema}"}
+      expect(connection.schema_search_path).to start_with %{"#{public_schema}"}
     end
   end
 
@@ -120,14 +120,14 @@ shared_examples_for "a schema based apartment adapter" do
     it "should reset connection" do
       subject.switch!(schema1)
       subject.reset
-      connection.schema_search_path.should start_with %{"#{public_schema}"}
+      expect(connection.schema_search_path).to start_with %{"#{public_schema}"}
     end
 
     context "with default_schema", :default_schema => true do
       it "should reset to the default schema" do
         subject.switch!(schema1)
         subject.reset
-        connection.schema_search_path.should start_with %{"#{default_schema}"}
+        expect(connection.schema_search_path).to start_with %{"#{default_schema}"}
       end
     end
 
@@ -138,13 +138,13 @@ shared_examples_for "a schema based apartment adapter" do
       end
 
       it "maintains the persistent schemas in the schema_search_path" do
-        connection.schema_search_path.should end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
+        expect(connection.schema_search_path).to end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
       end
 
       context "with default_schema", :default_schema => true do
         it "prioritizes the switched schema to front of schema_search_path" do
           subject.reset # need to re-call this as the default_schema wasn't set at the time that the above reset ran
-          connection.schema_search_path.should start_with %{"#{default_schema}"}
+          expect(connection.schema_search_path).to start_with %{"#{default_schema}"}
         end
       end
     end
@@ -153,12 +153,12 @@ shared_examples_for "a schema based apartment adapter" do
   describe "#switch!" do
     it "should connect to new schema" do
       subject.switch!(schema1)
-      connection.schema_search_path.should start_with %{"#{schema1}"}
+      expect(connection.schema_search_path).to start_with %{"#{schema1}"}
     end
 
     it "should reset connection if database is nil" do
       subject.switch!
-      connection.schema_search_path.should == %{"#{public_schema}"}
+      expect(connection.schema_search_path).to eq(%{"#{public_schema}"})
     end
 
     it "should raise an error if schema is invalid" do
@@ -176,7 +176,7 @@ shared_examples_for "a schema based apartment adapter" do
           subject.switch!(db)
         }.to_not raise_error
 
-        connection.schema_search_path.should start_with %{"#{db.to_s}"}
+        expect(connection.schema_search_path).to start_with %{"#{db.to_s}"}
       end
 
       after{ subject.drop(db) }
@@ -188,11 +188,11 @@ shared_examples_for "a schema based apartment adapter" do
       end
 
       it "should switch out the default schema rather than public" do
-        connection.schema_search_path.should_not include default_schema
+        expect(connection.schema_search_path).not_to include default_schema
       end
 
       it "should still switch to the switched schema" do
-        connection.schema_search_path.should start_with %{"#{schema1}"}
+        expect(connection.schema_search_path).to start_with %{"#{schema1}"}
       end
     end
 
@@ -201,11 +201,11 @@ shared_examples_for "a schema based apartment adapter" do
       before{ subject.switch!(schema1) }
 
       it "maintains the persistent schemas in the schema_search_path" do
-        connection.schema_search_path.should end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
+        expect(connection.schema_search_path).to end_with persistent_schemas.map { |schema| %{"#{schema}"} }.join(', ')
       end
 
       it "prioritizes the switched schema to front of schema_search_path" do
-        connection.schema_search_path.should start_with %{"#{schema1}"}
+        expect(connection.schema_search_path).to start_with %{"#{schema1}"}
       end
     end
   end
@@ -213,13 +213,13 @@ shared_examples_for "a schema based apartment adapter" do
   describe "#current" do
     it "should return the current schema name" do
       subject.switch!(schema1)
-      subject.current.should == schema1
+      expect(subject.current).to eq(schema1)
     end
 
     context "persistent_schemas", :persistent_schemas => true do
       it "should exlude persistent_schemas" do
         subject.switch!(schema1)
-        subject.current.should == schema1
+        expect(subject.current).to eq(schema1)
       end
     end
   end
