@@ -60,7 +60,7 @@ describe Apartment::Tenant do
     describe "#adapter" do
       it "should load postgresql adapter" do
         subject.adapter
-        Apartment::Adapters::PostgresqlAdapter.should be_a(Class)
+        expect(Apartment::Adapters::PostgresqlAdapter).to be_a(Class)
       end
 
       it "raises exception with invalid adapter specified" do
@@ -77,9 +77,9 @@ describe Apartment::Tenant do
 
         it 'has a threadsafe adapter' do
           subject.switch!(db1)
-          thread = Thread.new { subject.current.should == Apartment.default_tenant }
+          thread = Thread.new { expect(subject.current).to eq(Apartment.default_tenant) }
           thread.join
-          subject.current.should == db1
+          expect(subject.current).to eq(db1)
         end
       end
     end
@@ -100,7 +100,7 @@ describe Apartment::Tenant do
       describe "#create" do
         it "should seed data" do
           subject.switch! db1
-          User.count.should be > 0
+          expect(User.count).to be > 0
         end
       end
 
@@ -121,10 +121,10 @@ describe Apartment::Tenant do
             db_count = User.count + x.times{ User.create }
 
             subject.switch! db2
-            User.count.should == db2_count
+            expect(User.count).to eq(db2_count)
 
             subject.switch! db1
-            User.count.should == db_count
+            expect(User.count).to eq(db_count)
           end
         end
 
@@ -143,9 +143,9 @@ describe Apartment::Tenant do
 
             subject.switch! db1
             x.times{ Company.create }
-            Company.count.should == count + x
+            expect(Company.count).to eq(count + x)
             subject.reset
-            Company.count.should == count + x
+            expect(Company.count).to eq(count + x)
           end
         end
       end
@@ -165,8 +165,8 @@ describe Apartment::Tenant do
       it 'should seed from default path' do
         subject.create db1
         subject.switch! db1
-        User.count.should eq(3)
-        User.first.name.should eq('Some User 0')
+        expect(User.count).to eq(3)
+        expect(User.first.name).to eq('Some User 0')
       end
 
       it 'should seed from custom path' do
@@ -175,8 +175,8 @@ describe Apartment::Tenant do
         end
         subject.create db1
         subject.switch! db1
-        User.count.should eq(6)
-        User.first.name.should eq('Different User 0')
+        expect(User.count).to eq(6)
+        expect(User.first.name).to eq('Different User 0')
       end
     end
   end
