@@ -8,7 +8,7 @@ describe Apartment::Elevators::HostHash do
   describe "#parse_tenant_name" do
     it "parses the host for a domain name" do
       request = ActionDispatch::Request.new('HTTP_HOST' => 'example.com')
-      elevator.parse_tenant_name(request).should == 'example_tenant'
+      expect(elevator.parse_tenant_name(request)).to eq('example_tenant')
     end
 
     it "raises TenantNotFound exception if there is no host" do
@@ -24,7 +24,7 @@ describe Apartment::Elevators::HostHash do
 
   describe "#call" do
     it "switches to the proper tenant" do
-      Apartment::Tenant.should_receive(:switch).with('example_tenant')
+      expect(Apartment::Tenant).to receive(:switch).with('example_tenant')
 
       elevator.call('HTTP_HOST' => 'example.com')
     end
