@@ -1,3 +1,3 @@
 module Apartment
-  VERSION = "1.0.2"
+  VERSION = "1.2.0"
 end
